+++ scripts/session-start.js
@@ -333,6 +333,46 @@ async function upsertSession(db, s) {
     });
   }
 }
+var GREP_STOP_WORDS = new Set([
+  "a",
+  "an",
+  "and",
+  "are",
+  "as",
+  "at",
+  "be",
+  "but",
+  "by",
+  "do",
+  "does",
+  "for",
+  "from",
+  "have",
+  "how",
+  "i",
+  "in",
+  "is",
+  "it",
+  "of",
+  "on",
+  "or",
+  "our",
+  "such",
+  "that",
+  "the",
+  "this",
+  "to",
+  "use",
+  "was",
+  "we",
+  "what",
+  "when",
+  "where",
+  "which",
+  "why",
+  "with",
+  "you"
+]);
 async function snapshotFileHashes(db, projectId, hashes) {
   const now = Date.now();
   await db.batch(hashes.map((h) => ({